mgr/cephadm: move upgrade code to upgrade.py
Make it easier for everyone to read the upgrade algorithm.

(cherry picked from commit cedf9dbb2b29a787c92e38f2214cc9881a442d61)
